tcp: Fix reading into large buffers
Recent versions of Thunderbird limit the amount of data cached in a
socket for reading. Waiting until enough data is available to fill a
large buffer thus might cause the application to hang.

This commit thus alters the behavior of readFully to read out all
available data from the socket immediately, filling the given buffer in
multiple steps if required.