tcp: Move sockets into parent process (API change)
This commit moves sockets from the extension content process into the
parent process, which is necessary for upcoming versions of Thunderbird.
To reduce the performance impact, it also changes the API to no longer
permit unbounded read operations: read now takes a length and returns an
ArrayBuffer, readFully was removed.