customui: Basic support for Thunderbird Supernova
This commit introduces very basic support for changes in Thunderbird
leading up to the new UI "Supernova", scheduled to be released with
Thunderbird 115.

As of this commit, the experiment is operational with the new version;
some features like sidebar resizing are however not available when
running in a recent version of Thunderbird. Furthermore, there are a few
visual glitches regarding the locations using sidebar code changing
their desired size.