osfile: switch to IOUtils, move to parent
For some reasons, beta versions of Thunderbird 115 seem to have dropped
OS.File; this commit thus switches the osfile experiment to use the
successor API IOUtils. As Mozilla seems to tighten restrictions on child
processes and we don't care about performance at this point, this commit
furthermore moves the osfile experiment out of the child scope to prevent
future incompatibilities.